refactor: Reorganize components and add Footer
- Move layout components (Header, Logo) to dedicated layout directory
- Create minimalist Footer with logo and copyright
- Update import paths to use @/ alias pattern
- Ensure consistent theming across layout components
